refactor: コンポーネントに抽出
